rework rendertarget configuring

+++ Usagi/Runtime/Graphics/GpuImageCreateInfo.hpp
@@ -11,8 +11,8 @@ struct GpuImageCreateInfo
 {
     GpuBufferFormat format;
     Vector2u32 size;
-    std::uint32_t mip_levels = 1;
-    std::uint32_t sample_count = 1;
+    std::uint16_t mip_levels = 1;
+    std::uint16_t sample_count = 1;
     GpuImageUsage usage;
 };
 }
